+++ apps/frontman_server/lib/frontman_server/tasks/task.ex
@@ -15,14 +15,14 @@ defmodule FrontmanServer.Tasks.Task do
 
   use TypedStruct
 
-  alias FrontmanServer.Tasks.Execution.Framework
+  alias FrontmanServer.Frameworks
   alias FrontmanServer.Tasks.Interaction
 
   typedstruct enforce: true do
     field(:task_id, String.t())
     field(:short_desc, String.t())
     field(:interactions, list(Interaction.t()), default: [])
-    field(:framework, Framework.t())
+    field(:framework, Frameworks.t())
   end
 
   @doc """